Rename package and lambda handler class
